requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Limiter==3.12
Flask-CORS==5.0.0
PyJWT==2.10.1
marshmallow==3.26.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-cov==6.0.0
blinker>=1.9.0

